backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
openai
requests
python-dotenv
python-jose[cryptography]
